Add FlashStatusTracker to track status messages and update progress bar
Originally, the flash step printed status messages out to the command line,
which made sense because every step was happening serially. Now that there is
parallelization, we need a structured way to store the messages (except for
errors, which should interrupt the flash process) to read back later, e.g.
in logs (to be incorporated in a later commit).

Instead of displaying all the flash progress messages, now we display a
progress bar. Maybe we want a --verbose option but the end user should
mostly be concerned when errors occur, not with detailed status updates.

On that note information that we probably need to display to users here:
- List of all chips we are flashing
- What version we flashed from on each chip
- What version we flashed to on each chip
- If an update occurred or if ROM already matched on each chip
- When an error occurs, what chip it was associated with

Signed-off-by: Petra Alexson <palexson@tenstorrent.com>